+++ .github/workflows/crawling.yml
@@ -22,15 +22,15 @@ jobs:
           node-version: '22'
 
       - name: Install dependencies
-        run: npm install axios cheerio
+        run: npm install axios
 
       - name: Run Crawler
-        run: node getUserList.js # js 파일 실행
+        run: node getUserInfoList.js # js 파일 실행
 
       - name: Commit and Push changes
         run: |
           git config --global user.name "github-actions[bot]"
           git config --global user.email "github-actions[bot]@users.noreply.github.com"
-          git add -f userList.json
-          git commit -m "Update ranklist data" || echo "No changes to commit"
+          git add -f userInfoList.json
+          git commit -m "Update userInfoList data" || echo "No changes to commit"
           git push
