+++ userIdList.json
@@ -1,0 +1,3 @@
+{
+    "userIdList": ["dnjsgudtjr11", "cubic", "chunbae74", "castberry", "hjkiln", "soumatou", "gylim7826", "jjh4450", "dd2dd2", "kye04250425", "wschoi789", "qwer0987", "mule129", "glue0440", "hyun021218", "k2akr7", "mouiye", "kgh398", "binhan1e3", "parkha0623", "kihyjin153", "fcsolution11"]
+}
